fallout from ZuField/ZtField/ZfbField overhaul

+++ z/zfb/test/zfbtest3.cc
@@ -40,13 +40,13 @@ struct Order {
 
 ZfbFields(Order,
   (((symbol), (Keys<0>, Ctor<0>)), (String)),
-  (((orderID), (Keys<0>, Ctor<1>)), (UInt)),
+  (((orderID), (Keys<0>, Ctor<1>)), (UInt64)),
   (((link), ((Keys<1, 2>), Ctor<2>)), (String)),
   (((clOrdID), (Keys<1>, Ctor<3>)), (String)),
-  (((seqNo), (Keys<2>, Ctor<4>, Series, Index)), (UInt)),
+  (((seqNo), (Keys<2>, Ctor<4>, Series, Index)), (UInt64)),
   (((side), (Ctor<5>)), (Enum, Side::Map)),
-  (((price), (Ctor<6>)), (Int)),
-  (((quantity), (Ctor<7>)), (Int)));
+  (((price), (Ctor<6>)), (Int32)),
+  (((quantity), (Ctor<7>)), (Int32)));
 
 }
